+++ src/main/java/pam/pamhc2crops/items/ItemPamPig.java
@@ -32,7 +32,7 @@ public class ItemPamPig extends BlockNamedItem {
 
 		ItemStack stack = player.getHeldItem(hand);
 
-		if (!entity.world.isRemote && !entity.isChild() && entity instanceof AgeableEntity && (int) ((AgeableEntity) entity).getGrowingAge() == 0) {
+		if (!entity.world.isRemote && !entity.isChild() && entity instanceof AgeableEntity && ((AgeableEntity) entity).getGrowingAge() == 0) {
 			if (entity instanceof PigEntity) {
 				if (((PigEntity) entity).isInLove()) {
 					return ActionResultType.FAIL;
@@ -50,7 +50,7 @@ public class ItemPamPig extends BlockNamedItem {
 		if (entity.isChild()) {
 			if (!player.isCreative())
 				stack.shrink(1);
-			((AgeableEntity) entity).ageUp((int) ((float) (-((AgeableEntity) entity).getGrowingAge() / 20) * 0.1F),
+			((AgeableEntity) entity).ageUp((int) (-((AgeableEntity) entity).getGrowingAge() / 20 * 0.1F),
 					true);
 			return ActionResultType.PASS;
 		}
